trial.py has the code that will work at a basic level.
Still need to modify the distances between the nodes ie, the room.

As of approximate measurement the distance between navrt and navlt is 35m-36m.

The converter.py has the basic skeleton idea to represent the nodes graphically.

matrix_gen.py has the multiple codes to verify and ideate.

The main program has nothing of importance You can clear it off. The only reason it still there is for the libraries.